2024/09: Use a memory block approach
While solving I thought that maybe the simulation would be too slow so I
started a memory block approach. I messed it up big time:
- Store all in a HashMap (files and empty blocks combined)
- Deal with mutation in said HashMap (that was a disaster)

I was closer than I realized, because I had it working _except_ for
merging empty blocks. I thought this was required because when a file
moves it leaves an empty block that can be adjacent to another one.

However, files only move to the left, so the empty block left by the
file will never be used.

All things combined this cost me about 40 minutes and dead-ended. In
this commit is a solution that uses memory blocks the right way, with
the blocks stored in its own list, making everything a whole lot easier.